v0.8.1: - Switched from double precision to single precision - Replaced MTJ with jblas to support float matrices - Added FloatMath to apply basic functions from Java's Math class to floats - Added IntMath for some mathematical operations in the domain of integers - Replaced the custom tanh implementation with the function from the Math class - Added a convenience function to compute the byte size on the device - The only reference to the byte size constant Sizeof.FLOAT is now in this function. - Removed default values in optimizer and instruction constructors - Replaced some occurrences of integers with floats to avoid implicit conversions - Replaced RMSProp with Momentum in the TREC demo: the new model achieves 92.8% accuracy after 10 iterations - Refactored result access in cuBLAS projection and CUDA stochastic gradient descent tests

+++ src/main/kotlin/shape/komputation/cpu/demos/mnist/MnistDropoutNesterov.kt
@@ -35,9 +35,9 @@ fun main(args: Array<String>) {
     val hiddenDimension = 100
     val numberCategories = MnistData.numberCategories
 
-    val initialization = gaussianInitialization(random, 0.0, 0.1)
+    val initialization = gaussianInitialization(random, 0.0f, 0.1f)
 
-    val optimizer = nesterov(0.005, 0.1)
+    val optimizer = nesterov(0.005f, 0.1f)
 
     val firstProjection = projectionLayer(
         inputDimension,
@@ -59,11 +59,11 @@ fun main(args: Array<String>) {
     val network = Network(
         inputLayer(inputDimension),
         firstProjection,
-        dropoutLayer(hiddenDimension, random, 0.85, reluLayer(hiddenDimension)),
+        dropoutLayer(hiddenDimension, random, 0.85f, reluLayer(hiddenDimension)),
         outputLayer
     )
 
-    val afterEachIteration = { _ : Int, _ : Double ->
+    val afterEachIteration = { _ : Int, _ : Float ->
 
         val accuracy = network
             .test(
@@ -76,7 +76,8 @@ fun main(args: Array<String>) {
                 }
             )
             .count { correct -> correct }
-            .div(MnistData.numberTestExamples.toDouble())
+            .toFloat()
+            .div(MnistData.numberTestExamples.toFloat())
 
         println(accuracy)
