+++ src/main/kotlin/shape/komputation/cpu/layers/SparseForwarding.kt
@@ -1,9 +1,9 @@
 package shape.komputation.cpu.layers
 
-import shape.komputation.matrix.DoubleMatrix
+import shape.komputation.matrix.FloatMatrix
 
 interface SparseForwarding {
 
-    fun forward(input: DoubleMatrix, mask: BooleanArray): DoubleMatrix
+    fun forward(input: FloatMatrix, mask: BooleanArray): FloatMatrix
 
 }
