+++ src/main/kotlin/shape/komputation/cpu/layers/combination/HadamardCombination.kt
@@ -2,31 +2,31 @@ package shape.komputation.cpu.layers.combination
 
 import shape.komputation.cpu.functions.hadamard
 import shape.komputation.cpu.layers.CombinationLayer
-import shape.komputation.matrix.DoubleMatrix
+import shape.komputation.matrix.FloatMatrix
 
 class HadamardCombination internal constructor(val name: String?) : CombinationLayer(name) {
 
-    private var first = DoubleArray(0)
-    private var second = DoubleArray(0)
+    private var first = FloatArray(0)
+    private var second = FloatArray(0)
 
-    override fun forward(first: DoubleMatrix, second: DoubleMatrix): DoubleMatrix {
+    override fun forward(first: FloatMatrix, second: FloatMatrix): FloatMatrix {
 
         this.first = first.entries
         this.second = second.entries
 
-        return DoubleMatrix(first.numberRows, first.numberColumns, hadamard(first.entries, second.entries))
+        return FloatMatrix(first.numberRows, first.numberColumns, hadamard(first.entries, second.entries))
 
     }
 
     // d f(x) * g(x) / d f(x) = g(x)
-    override fun backwardFirst(chain: DoubleMatrix) =
+    override fun backwardFirst(chain: FloatMatrix) =
 
-        DoubleMatrix(chain.numberRows, chain.numberColumns, hadamard(chain.entries, this.second))
+        FloatMatrix(chain.numberRows, chain.numberColumns, hadamard(chain.entries, this.second))
 
     // d f(x) * g(x) / d g(x) = f(x)
-    override fun backwardSecond(chain: DoubleMatrix) =
+    override fun backwardSecond(chain: FloatMatrix) =
 
-        DoubleMatrix(chain.numberRows, chain.numberColumns, hadamard(chain.entries, this.first))
+        FloatMatrix(chain.numberRows, chain.numberColumns, hadamard(chain.entries, this.first))
 
 }
